Identity ekle methodlarında bigint donduruyor. Arkadas cast ediyor. gerekli int ve short, byte'a
git-svn-id: https://karkas.googlecode.com/svn/trunk@548 ac15386d-a63e-0410-a15c-c17449491352

+++ Karkas.Core/Karkas.Core.DataUtil/BaseBs.cs
@@ -109,7 +109,7 @@ namespace Karkas.Core.DataUtil
 
 
         protected DAL_TIPI dal = new DAL_TIPI();
-        public int Ekle(TYPE_LIBRARY_TIPI k)
+        public long Ekle(TYPE_LIBRARY_TIPI k)
         {
             return dal.Ekle(k);
         }
